Fix compilation errors

+++ common/src/thrift/tachyon.thrift
@@ -44,8 +44,9 @@ struct FileBlockInfo {
   4: list<NetAddress> locations
 }
 
+// TODO: make fileId into i64
 struct FileInfo {
-  1: i64 fileId
+  1: i32 fileId
   2: string name
   3: string path
   4: string ufsPath
